start day 6

+++ advent-of-code-2021/src/main/kotlin/aoc/Day06LanternFish.kt
@@ -1,18 +1,18 @@
 package aoc
 
 import java.io.File
-import java.math.BigInteger
+import java.math.BigInteger as BigInt
 
 internal object Day06LanternFish : ChallengeDay {
 
     private const val INIT_TIMER_VAL_NEW_BORN = 8
     private const val TIMER_VAL_AFTER_SPAWN = 6
 
-    fun part1(path: String): BigInteger = File(path).toFishCount(days = 80)
+    fun part1(path: String): BigInt = File(path).toFishCount(days = 80)
 
-    fun part2(path: String): BigInteger = File(path).toFishCount(days = 256)
+    fun part2(path: String): BigInt = File(path).toFishCount(days = 256)
 
-    private fun File.toFishCount(days: Int): BigInteger {
+    private fun File.toFishCount(days: Int): BigInt {
         val daysLeftTillNewSpawnCounts = toCountsArray()
         for (day in 1..days) {
             val curDaysLeftTillNewSpawnCounts = daysLeftTillNewSpawnCounts.copyOf()
@@ -27,8 +27,8 @@ internal object Day06LanternFish : ChallengeDay {
         return daysLeftTillNewSpawnCounts.sumOf { it }
     }
 
-    private fun File.toCountsArray(): Array<BigInteger> {
-        val daysLeftTillNewSpawn = Array(INIT_TIMER_VAL_NEW_BORN + 1) { BigInteger.ZERO }
+    private fun File.toCountsArray(): Array<BigInt> {
+        val daysLeftTillNewSpawn = Array(INIT_TIMER_VAL_NEW_BORN + 1) { BigInt.ZERO }
         readText().split(',').map(String::trim).map(String::toInt)
             .forEach { daysLeft -> daysLeftTillNewSpawn[daysLeft]++ }
         return daysLeftTillNewSpawn
